Add default answer in case there is no answer in the SO question.
There are cases, e.g. `howdoi -n3 run an http server in python`, where there are
answers in the search results which are empty (e.g. there's a SO question, but
it has no answers). Currently, howdoi crashes on these. Ideally, I should
probably check to see if a link has an answer, and toss out those which do not.
This is a little fix though, and prevents the program from crashing.