Remove unnecessary test code that supported Python 2.6
When writing automated tests for #177, I learned that
Python 2.6 did not have support for assertSequenceEqual. To ensure
that the automated tests could run on Python 2.6, I wrote some
additional code.

However, in #186, we dropped support for Python 2.6. Therefore,
we no longer need this code. Removing it would help simplify
the test code and make it easier to maintain.