Display no link-associated markdown if the copy duplicates the href
Sometimes a user may paste a URL directly into an answer
and Stack Overflow formats that URL as a valid hyperlink.
This means that if we try to convert every hyperlink
into markdown, we will get output like
[https://example.com](https://example.com), which is repetitive
and useless.

Therefore, before we do any replacement, we check if the href
and the copy are equal to each other. If they are, then we don't
provide a Markdown version of the hyperlink. Instead, we simply
use the plain text version.