Use better query for selecting the correct answer from SO
Previously we would look for the correct answer inside the `answer`
class. This div actually contains the answer _and_ the comments. It's
a better user experience to instead use the `answercell` class which
only contains the answer and not the comments.

Also added a test for this use-case, and refreshed all the old tests
by deleting and letting `nosetest` recreate the `page_cache` directory.